mmpose/requirements/runtime.txt
chumpy
json_tricks
matplotlib
munkres
numpy
opencv-python
pillow
scipy
torchvision
xtcocotools>=1.12
tomesd
